Final touches and add cost details to user purchase list

+++ user_login/models.py
@@ -7,7 +7,7 @@ class User_Purchase_List(models.Model):
     user_email = models.CharField(max_length=50, unique=False)
     company_phone = models.CharField(max_length=30, unique=False)
     company_email = models.CharField(max_length=50, unique=True)
-    ad_price = models.IntegerField(unique=False)
+    ad_price = models.FloatField(unique=False)
     ad_name = models.CharField(max_length=255, unique=False)
     purchase_date  = models.DateField(unique=False)
     rating = models.IntegerField(null = True, unique=False)
